% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{bayes_lm}
\alias{bayes_lm}
\title{Bayesian Linear Model}
\usage{
bayes_lm(
  y,
  x,
  ...,
  iterations = 10000,
  mean_prior = 1,
  beta_prior = NULL,
  variance_prior = NULL
)
}
\arguments{
\item{y}{A vector of response values}

\item{x}{A matrix of input values}

\item{...}{Placeholder for future capabilities. Currently ignored.}

\item{iterations}{The number of iterations in the Gibbs Sampler}

\item{mean_prior}{The prior value for the mean}

\item{beta_prior}{The prior values for the regression coefficients}

\item{variance_prior}{The prior value for variance}
}
\value{
A matrix of posterior values where each column is a parameter. The
regression coefficients are first, then the variance.
}
\description{
\code{bayes_lm()} uses a Gibbs Sampler to create a Bayesian linear model.
}
\examples{
# Simulate data
X <- matrix(rnorm(1000, 5, 2), nrow = 100)
betas <- runif(10, -2, 2)
Y <- X \%*\% betas + rnorm(100, 0, 3)

# Perform Bayesian Linear Regression
bayes_lm(Y, X, iterations = 10)
}
