% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{coefficients_post}
\alias{coefficients_post}
\title{Compute Posterior Regression Coefficients}
\usage{
coefficients_post(X, iters, G, b_prior, v_post)
}
\arguments{
\item{X}{A matrix of input values}

\item{iters}{The number of iterations to perform}

\item{G}{The unit information prior. Equivalent to the number of rows in X}

\item{b_prior}{The prior values of the regression coefficients}

\item{v_post}{The posterior values for the variance}
}
\value{
A matrix of regression coefficients where each column represents an
iteration in the Gibbs Sampler.
}
\description{
\code{coefficients_post()} computes the posterior regression coefficients for the
linear regression model
}
