% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{credible_intervals}
\alias{credible_intervals}
\title{Linear Regression Credible Intervals}
\usage{
credible_intervals(posterior_values, ..., level = 0.9)
}
\arguments{
\item{posterior_values}{The result from \code{bayes_lm()}}

\item{...}{Placeholder for future capabilities. Currently ignored}

\item{level}{The level for the interval}
}
\value{
A matrix showing the credible intervals
}
\description{
\code{credible_intervals()} computes the credible interval for each parameter in a
model created by \code{bayes_lm()}.
}
\examples{
# Simulate data
X <- matrix(rnorm(1000, 5, 2), nrow = 100)
betas <- runif(10, -2, 2)
Y <- X \%*\% betas + rnorm(100, 0, 3)

# Perform Bayesian Linear Regression
linear_model <- bayes_lm(Y, X, iterations = 10)

# Compute credible intervals
credible_intervals(linear_model)
}
