% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_parameter.R
\name{dbeta_post}
\alias{dbeta_post}
\title{Sample Posterior Beta PDF}
\usage{
dbeta_post(
  quanitle,
  successes,
  sample_size,
  ...,
  alpha_prior = 1/2,
  beta_prior = 1/2
)
}
\arguments{
\item{successes}{number of success in the binomial experiment}

\item{sample_size}{number of trials in the binomial experiment}

\item{...}{Placeholder for future capabilities. Currently ignored.}

\item{alpha_prior}{The prior for the alpha parameter}

\item{beta_prior}{The prior for the beta parameter}

\item{quantile}{vector of quantiles}
}
\value{
A vector of probabilities the same length as quantile
}
\description{
\code{dbeta_post()} will sample the posterior beta probability distribution function
in a beta-binomial model.
}
