% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_priors.R
\name{dgamma_post}
\alias{dgamma_post}
\title{Posterior Gamma Distribution}
\usage{
dgamma_post(
  quantile,
  observations,
  sample_size,
  likelihood,
  ...,
  alpha_observed = NULL,
  alpha_prior = 0,
  beta_prior = 0
)
}
\arguments{
\item{quantile}{A vector of quantiles}

\item{observations}{The observations from the model}

\item{sample_size}{The sample size from the model}

\item{likelihood}{The likelihood used}

\item{...}{Pass arguments to dgamma}

\item{alpha_observed}{The observed shape parameter}

\item{alpha_prior}{The prior shape parameter}

\item{beta_prior}{The prior rate parameter}
}
\value{
A vector of PDF results that is the same length as quantile.
}
\description{
\code{dgamma_post()} computes the posterior PDF for a model with a gamma prior.
}
\details{
\code{pgamma_post()} computes the posterior CDF for a model with a gamma prior.

\code{rgamma_post()} generates random numbers from the posterior distribution from
a model with a gamma prior.
}
