% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_priors.R
\name{dnorm_post}
\alias{dnorm_post}
\alias{pnorm_post}
\alias{qnorm_post}
\alias{rnorm_post}
\title{Posterior Normal Distribution}
\usage{
dnorm_post(
  quantiles,
  observations,
  ...,
  mean_prior = 1,
  variance_prior = 1/stats::sd(observations),
  precision_prior = 1/variance_prior
)

pnorm_post(
  quantiles,
  observations,
  ...,
  mean_prior = 1,
  variance_prior = 1/stats::sd(observations),
  precision_prior = 1/variance_prior
)

qnorm_post(
  percentiles,
  observations,
  ...,
  mean_prior = 1,
  variance_prior = 1/stats::sd(observations),
  precision_prior = 1/variance_prior
)

rnorm_post(
  n,
  observations,
  ...,
  mean_prior = 1,
  variance_prior = 1/stats::sd(observations),
  precision_prior = 1/variance_prior
)
}
\arguments{
\item{quantiles}{A vector of quantiles}

\item{observations}{The observations from the model}

\item{...}{Pass parameters to the underlying functions such as log or
lower.tail}

\item{mean_prior}{The prior value for the mean}

\item{variance_prior}{The prior value for the variance}

\item{precision_prior}{The prior value for the variance}

\item{percentiles}{A vector of percentiles}

\item{n}{The number of random values to generate}
}
\value{
A vector the same length as \code{quantiles} or \code{n}
}
\description{
\code{dnorm_post()} computes values from the posterior PDF for a model with a
normal prior.
}
\details{
\code{pnorm_post()} computes values from the posterior CDF for a model with a
normal prior.

\code{qnorm_post()} computes the quantile corresponding with a percentile for a
model with a normal prior.

\code{rnorm_post} generates a random value from the posterior distribution for a
model with a normal prior.

These posterior functions assume that the variance for the model is fixed.
}
\examples{
x <- seq(0, 1, by = 0.1)
obs <- rnorm(50)
dnorm_post(x, obs)
pnorm_post(x, obs)
qnorm_post(seq(0.1, 0.9, by = 0.1), obs)

# Random numbers can also be generated
rnorm_post(5, obs)
}
