% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{initialize_priors}
\alias{initialize_priors}
\title{Initialize linear model priors}
\usage{
initialize_priors(Y, X, g_frac, b_prior, v_prior)
}
\arguments{
\item{Y}{A vector of response values}

\item{X}{A matrix of input values}

\item{g_frac}{The unit information divided by itself plus 1}

\item{b_prior}{The prior values for beta}

\item{v_prior}{The prior values for the variance}
}
\value{
A list of the priors. The beta priors are first then the variance.
}
\description{
\code{initialize_priors()} ensures that both beta and the variance have priors.
}
