% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_parameter.R
\name{pbeta_post}
\alias{pbeta_post}
\title{Sample from beta CDF}
\usage{
pbeta_post(
  quantile,
  successes,
  sample_size,
  ...,
  alpha_prior = 1/2,
  beta_prior = 1/2
)
}
\arguments{
\item{quantile}{Vector of quantiles}

\item{successes}{The number of successes in the binomial experiment}

\item{sample_size}{The sample size for the binomial experiment}

\item{...}{Pass additional arguments to the pbeta function}

\item{alpha_prior}{The prior value for alpha}

\item{beta_prior}{The prior value for beta}
}
\value{
A vector of probabilities the same length as the quantiles
}
\description{
Samples from the Cumulative Density Function for the beta distribution.
}
