% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplebayes-package.R
\docType{package}
\name{simplebayes-package}
\alias{simplebayes}
\alias{simplebayes-package}
\title{simplebayes: Implement Simple Bayesian Statistics}
\description{
Compute posterior distributions from simple Bayesian Models. This can be done for models with Beta, Gamma, and Normal priors as well as for a linear model.
}
\author{
\strong{Maintainer}: Kevin Green \email{greenk12@tamu.edu}

}
\keyword{internal}
