% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplots.R
\name{traceplot}
\alias{traceplot}
\title{Create Trace Plots}
\usage{
traceplot(bayes_lm_distributions)
}
\arguments{
\item{bayes_lm_distributions}{The result from \code{bayes_lm}}
}
\value{
A ggplot2 plot of trace plots
}
\description{
\code{traceplot()} creates the trace plots for the results from \code{bayes_lm()}. This
function requires the ggplot2 and tidyr packages from the tidyverse.
}
\examples{
#' # Simulate data
X <- matrix(rnorm(1000, 5, 2), nrow = 100)
betas <- runif(10, -2, 2)
Y <- X \%*\% betas + rnorm(100, 0, 3)

# Perform Bayesian Linear Regression
result <- bayes_lm(Y, X, iterations = 10)

# Plot data
traceplot(result)
}
