% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_priors.R
\name{update_beta_parameters}
\alias{update_beta_parameters}
\title{Compute Beta Posterior Parameters}
\usage{
update_beta_parameters(
  a_prior,
  b_prior,
  l_hood,
  res = NULL,
  s_size = NULL,
  f_rate = NULL
)
}
\arguments{
\item{a_prior}{The alpha prior value}

\item{b_prior}{The beta prior value}

\item{l_hood}{The likelihood for the model}

\item{res}{The result of the experiment (either successes or failures)}

\item{s_size}{The sample size of the experiment}

\item{f_rate}{The failure rate for the model. Only used for the negative
binomial likelihood.}
}
\value{
A vector of length two where the first value is the alpha parameter
and the second value is the beta parameter.
}
\description{
\code{update_beta_parameters()} will compute the posterior parameter values
for a model with a beta prior.
}
