% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_priors.R
\name{update_gamma_parameters}
\alias{update_gamma_parameters}
\title{Update Gamma Parameters}
\usage{
update_gamma_parameters(
  a_prior,
  b_prior,
  l_hood,
  ...,
  obs = NULL,
  s_size = NULL,
  a_obs = NULL
)
}
\arguments{
\item{a_prior}{The shape parameter}

\item{b_prior}{The rate parameter}

\item{l_hood}{The likelihood of the model}

\item{...}{Placeholder for future capabilities. Currently ignored}

\item{obs}{The observations from the experiment}

\item{s_size}{The sample size from the experiment}

\item{a_obs}{The observed shape parameter}
}
\value{
A vector of length two where the shape parameter is first and then
the rate parameter
}
\description{
\code{update_gamma_parameters()} will update the parameters for a gamma prior for
a given likelihood. Currently, these are the Poisson, exponential, gamma, and
inverse gamma likelihoods.
}
