% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_priors.R
\name{update_normal_parameters}
\alias{update_normal_parameters}
\title{Update Normal Parameters}
\usage{
update_normal_parameters(
  mean_prior,
  var_prior,
  obs,
  ...,
  precision_prior = 1/var_prior
)
}
\arguments{
\item{mean_prior}{The prior value for the mean}

\item{var_prior}{The prior value for the variance}

\item{obs}{The observations from the model}

\item{...}{Placeholder for future capabilities. Currently ignored}

\item{precision_prior}{The prior value for the precision}
}
\value{
A vector of length 2 where the posterior mean is first followed by
the posterior variance
}
\description{
Update Normal Parameters
}
