% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{variance_post}
\alias{variance_post}
\title{Compute Posterior Variance}
\usage{
variance_post(Y, X, iters, g, mu_prior, s2_prior)
}
\arguments{
\item{Y}{A vector of response values}

\item{X}{A matrix of input values}

\item{iters}{The number of iterations to perform}

\item{g}{The unit information. Equivalent to the length of Y}

\item{mu_prior}{The prior for mu}

\item{s2_prior}{The prior for the variance}
}
\value{
A vector of posterior variance values
}
\description{
\code{variance_post()} computes the posterior variance for the linear model.
}
