% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_priors.R
\name{dbeta_post}
\alias{dbeta_post}
\alias{pbeta_post}
\alias{qbeta_post}
\alias{rbeta_post}
\title{Posterior Beta Distribution}
\usage{
dbeta_post(
  quantile,
  result,
  sample_size,
  likelihood,
  ...,
  failure_rate = NULL,
  alpha_prior = 0.5,
  beta_prior = 0.5
)

pbeta_post(
  quantile,
  result,
  sample_size,
  likelihood,
  ...,
  failure_rate = NULL,
  alpha_prior = 0.5,
  beta_prior = 0.5
)

qbeta_post(
  percentile,
  result,
  sample_size,
  likelihood,
  ...,
  failure_rate = NULL,
  alpha_prior = 0.5,
  beta_prior = 0.5
)

rbeta_post(
  n,
  result,
  sample_size,
  likelihood,
  ...,
  failure_rate = NULL,
  alpha_prior = 0.5,
  beta_prior = 0.5
)
}
\arguments{
\item{quantile}{A vector of quantiles}

\item{result}{The result from the experiment as successes or failures}

\item{sample_size}{The sample size from the experiement}

\item{likelihood}{The likelihood to be used}

\item{...}{Pass arguments to the underlying xbeta functions.}

\item{failure_rate}{The failure rate of a negative binomial experiement}

\item{alpha_prior}{The prior value for the alpha parameter}

\item{beta_prior}{The prior value for the beta parameter}

\item{percentile}{A vector of percentiles}

\item{n}{The number of random values to generate}
}
\value{
A vector of the same length as \code{quantile}.
}
\description{
\code{dbeta_post()} will compute the posterior CDF of a model with a beta
prior. Currently, the binomial, negative binomial, and geometric likelihoods
can be used.
}
\details{
\code{pbeta_post()} will compute the posterior PDF of a model with a beta
prior.

\code{qbeta_post()} computes the quantile for a given probability from a model
with a beta prior.

\code{rbeta_post()} will generate random numbers from the posterior distribution
of a model with a beta prior.
}
\examples{

x <- seq(0, 1, by = 0.1)
dbeta_post(x, 5, 10, "binom")
pbeta_post(x, 5, 10, "binom")
qbeta_post(seq(0.1, 0.9, by = 0.1), 5, 10, "binom")

# Can also generate random numbers
rbeta_post(3, 5, 10, "binom")
}
