% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_priors.R
\name{dgamma_post}
\alias{dgamma_post}
\alias{pgamma_post}
\alias{qgamma_post}
\alias{rgamma_post}
\title{Posterior Gamma Distribution}
\usage{
dgamma_post(
  quantile,
  observations,
  sample_size,
  likelihood,
  ...,
  alpha_observed = NULL,
  alpha_prior = 0,
  beta_prior = 0
)

pgamma_post(
  quantile,
  observations,
  sample_size,
  likelihood,
  ...,
  alpha_observed = NULL,
  alpha_prior = 0,
  beta_prior = 0
)

qgamma_post(
  percentile,
  observations,
  sample_size,
  likelihood,
  ...,
  alpha_observed = NULL,
  alpha_prior = 0,
  beta_prior = 0
)

rgamma_post(
  n,
  observations,
  sample_size,
  likelihood,
  ...,
  alpha_observed = NULL,
  alpha_prior = 0,
  beta_prior = 0
)
}
\arguments{
\item{quantile}{A vector of quantiles}

\item{observations}{The observations from the model}

\item{sample_size}{The sample size from the model}

\item{likelihood}{The likelihood used}

\item{...}{Pass arguments to dgamma}

\item{alpha_observed}{The observed shape parameter}

\item{alpha_prior}{The prior shape parameter}

\item{beta_prior}{The prior rate parameter}

\item{n}{The number of random values to generate}
}
\value{
A vector of PDF results that is the same length as quantile.
}
\description{
\code{dgamma_post()} computes the posterior PDF for a model with a gamma prior.
}
\details{
\code{pgamma_post()} computes the posterior CDF for a model with a gamma prior.

\code{qgamma_post()} computes the quantile for a given probability from a model
with a gamma prior.

\code{rgamma_post()} generates random numbers from the posterior distribution from
a model with a gamma prior.
}
\examples{

x <- seq(0, 1, by = 0.1)
dgamma_post(x, 10, 100, "pois")
pgamma_post(x, 10, 100, "pois")
qgamma_post(x, 10, 100, "pois")

# Random numbers can also be generated
rgamma_post(5, 10, 100, "pois")
}
