% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplebayes-package.R
\docType{package}
\name{simplebayes-package}
\alias{simplebayes}
\alias{simplebayes-package}
\title{simplebayes: Implement Simple Bayesian Statistics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kevin Green \email{greenk12@tamu.edu}

}
\keyword{internal}
